+++ codelite_terminal/MainFrame.cpp
@@ -1,21 +1,21 @@
 #include "MainFrame.h"
-#include <wx/aboutdlg.h>
-#include <wx/regex.h>
-#include <errno.h>
-#include <wx/settings.h>
-#include <wx/colordlg.h>
 #include "SettingsDlg.h"
-#include <wx/filedlg.h>
+#include <errno.h>
+#include <wx/aboutdlg.h>
 #include <wx/clipbrd.h>
+#include <wx/colordlg.h>
 #include <wx/ffile.h>
+#include <wx/filedlg.h>
 #include <wx/filename.h>
+#include <wx/regex.h>
+#include <wx/settings.h>
 
 #ifndef __WXMSW__
 #if defined(__FreeBSD__)
-#include <sys/types.h>
-#include <sys/ioctl.h>
-#include <termios.h>
 #include <libutil.h>
+#include <sys/ioctl.h>
+#include <sys/types.h>
+#include <termios.h>
 #elif defined(__WXGTK__)
 #include <pty.h>
 #else
@@ -62,13 +62,12 @@ MainFrame::MainFrame(wxWindow* parent, const TerminalOptions& options, long styl
     m_stc->SetFont(wxSystemSettings::GetFont(wxSYS_SYSTEM_FIXED_FONT));
     wxString tty = StartTTY();
     SetCartAtEnd();
-    m_stc->MarkerDefine(MARKER_ID, wxSTC_MARK_ARROWS);
-    m_stc->MarkerSetBackground(MARKER_ID, *wxBLACK);
+
     // m_stc->MarkerSetAlpha(MARKER_ID, 5);
     SetSize(m_config.GetTerminalSize());
     SetPosition(m_config.GetTerminalPosition());
 
-    DoApplySettings();
+    // DoApplySettings();
     CallAfter(&MainFrame::DoExecStartCommand);
 
     // Connect color menu items
@@ -78,6 +77,7 @@ MainFrame::MainFrame(wxWindow* parent, const TerminalOptions& options, long styl
     Connect(ID_SIGINT, wxEVT_COMMAND_MENU_SELECTED, wxCommandEventHandler(MainFrame::OnSignal), NULL, this);
     Connect(ID_SIGTERM, wxEVT_COMMAND_MENU_SELECTED, wxCommandEventHandler(MainFrame::OnSignal), NULL, this);
     Connect(ID_SIGHUP, wxEVT_COMMAND_MENU_SELECTED, wxCommandEventHandler(MainFrame::OnSignal), NULL, this);
+    DoApplySettings();
 }
 
 MainFrame::~MainFrame()
@@ -153,9 +153,7 @@ void MainFrame::OnKeyDown(wxKeyEvent& event)
         event.Skip();
 
     } else if(event.GetKeyCode() == WXK_BACK) {
-        if((m_stc->GetCurrentPos() - 1) < m_fromPos) {
-            return;
-        }
+        if((m_stc->GetCurrentPos() - 1) < m_fromPos) { return; }
         event.Skip();
 
     } else {
@@ -174,9 +172,7 @@ void MainFrame::DoExecuteCurrentLine(const wxString& command)
         async = true;
     }
 
-    if(command.IsEmpty()) {
-        AppendNewLine();
-    }
+    if(command.IsEmpty()) { AppendNewLine(); }
 
     if(cmd.IsEmpty()) {
         SetCartAtEnd();
@@ -188,7 +184,7 @@ void MainFrame::DoExecuteCurrentLine(const wxString& command)
     if(reCD.Matches(cmd)) {
         reCD.Replace(&cmd, "");
         if(::wxSetWorkingDirectory(cmd)) {
-            m_stc->AppendText("current directory: " + ::wxGetCwd() + "\n");
+            m_stc->AppendText(::wxGetCwd() + "\n");
 
         } else {
             m_stc->AppendText(wxString(strerror(errno)) + "\n");
@@ -322,9 +318,7 @@ void MainFrame::OnClearView(wxCommandEvent& event)
 
 void MainFrame::OnTerminateInfirior(wxCommandEvent& event)
 {
-    if(m_process) {
-        m_process->Terminate();
-    }
+    if(m_process) { m_process->Terminate(); }
 }
 
 void MainFrame::OnTerminateInfiriorUI(wxUpdateUIEvent& event) { event.Enable(m_process != NULL); }
@@ -402,24 +396,15 @@ void MainFrame::OnSettings(wxCommandEvent& event)
 
 void MainFrame::DoApplySettings()
 {
-    DoSetColour(m_config.GetFgColour(), false);
-    DoSetColour(m_config.GetBgColour(), true);
-    DoSetFont(m_config.GetFont());
+    m_stc->SetPreferences(m_config.GetFont(), m_config.GetFgColour(), m_config.GetBgColour());
 }
 
 void MainFrame::OnSaveContent(wxCommandEvent& event)
 {
     const wxString ALL(wxT("All Files (*)|*"));
-    wxFileDialog dlg(this,
-                     _("Save As"),
-                     ::wxGetCwd(),
-                     "codelite-terminal.txt",
-                     ALL,
-                     wxFD_SAVE | wxFD_OVERWRITE_PROMPT,
+    wxFileDialog dlg(this, _("Save As"), ::wxGetCwd(), "codelite-terminal.txt", ALL, wxFD_SAVE | wxFD_OVERWRITE_PROMPT,
                      wxDefaultPosition);
-    if(dlg.ShowModal() == wxID_OK) {
-        m_stc->SaveFile(dlg.GetPath());
-    }
+    if(dlg.ShowModal() == wxID_OK) { m_stc->SaveFile(dlg.GetPath()); }
 }
 
 void MainFrame::OnSaveContentUI(wxUpdateUIEvent& event) { event.Enable(!m_stc->IsEmpty()); }
@@ -435,9 +420,7 @@ void MainFrame::OnIdle(wxIdleEvent& event)
 void MainFrame::AppendOutputText(const wxString& text)
 {
     m_outoutBuffer << text;
-    if(m_outoutBuffer.length() > wxMEGA_BYTE) {
-        FlushOutputBuffer();
-    }
+    if(m_outoutBuffer.length() > wxMEGA_BYTE) { FlushOutputBuffer(); }
 }
 
 void MainFrame::FlushOutputBuffer()
